requirements.txt
CairoSVG
Flask
Flask-CacheBuster
Flask-Cors
Flask-Login
IMAPClient
Jinja2
MarkupSafe
Pillow
PyMuPDF
PyPDF2
PyWavelets
PyYAML
Pygments
Shapely
SoundFile
SpeechRecognition
Wand
Werkzeug
XlsxWriter
absl-py
analytics-python
attrs
audioread
beautifulsoup4
bokeh
cachetools
cairosvg
chardet
click
click-plugins
compressed-rtf
debugpy
decorator
defusedxml
deprecat
docx2txt
ebooklib
email-validator
fastapi
fastjsonschema
ffmpeg-python
ffmpy
fire
fitz
flask
flask-cachebuster
flask-cors
flask-login
fonttools
frontend
fpdf
future
fuzzywuzzy
gensim==3.8.2
gradio
graphviz
h5py
html5lib
httpx
imageio
imageio-ffmpeg
imgkit
ipython
jinja2
jieba
json5
jsonpickle
jsonschema
jupyter-client
jupyter-core
jupyter-server
jupyterlab
jupyterlab-pygments
jupyterlab-server
keras
langchain
langchain-experimental
librosa
loguru
lxml
markdown2
markdownify
matplotlib
matplotlib-inline
matplotlib-venn
moviepy
murmurhash
nbclient
nbconvert
nbformat
networkx
nltk
notebook
numba
numexpr
numpy
numpy-financial
openai
opencv-python
openpyxl
orjson
pandas
pdf2image
pdfplumber
pdfkit
pdfminer.six
pdfplumber
pillow
plotly
psutil
pydantic
PyMuPDF
pydub
pygments
pylog
pymupdf
pypandoc
pyparsing
pypdf2==2.12.0
pytesseract
pytest
python-dateutil
python-docx
python-multipart
python-pptx
pytz
pywavelets
pygame
pyyaml
qrcode
rarfile
requests
scikit-image
scikit-learn
scipy
seaborn
sentencepiece
soundfile
speechrecognition
starlette
statsmodels
svglib
svgwrite
sympy
sumy
tabula
tabulate
textblob
tifffile
toml
torch
torchaudio
torchtext
torchvision
tornado
tqdm
typing-extensions
tzdata
tzlocal
ujson
urllib3
uvicorn
wand
websocket-client
websockets
werkzeug
wordcloud
xgboost
xlrd
xlsxwriter
xml-python
zipptextblob
audioread
librosa
PyAudio
pydub
fbprophet